refactor(core): declare the value recorder once per feature
ConditionRewriter previously declared $spock_valueRecorder and
$spock_errorCollector for every condition-bearing block, producing
duplicate declarations in features with conditions in multiple blocks.
Hoist their creation to FeatureRewriter so they are declared once per
feature and shared across all blocks, matching Spock. isConditionStatement
is extracted as a shared helper so the feature rewriter can decide up
front whether a recorder is needed.
