+++ spockk-compiler-plugin/src/main/kotlin/io/github/pshevche/spockk/compilation/transformer/condition/ConditionRewriter.kt
@@ -14,36 +14,26 @@
 
 package io.github.pshevche.spockk.compilation.transformer.condition
 
-import io.github.pshevche.spockk.compilation.ir.IrIdentifiers
-import io.github.pshevche.spockk.compilation.ir.findFieldByName
-import io.github.pshevche.spockk.compilation.ir.findRequiredClassSymbol
 import io.github.pshevche.spockk.compilation.ir.irCatchParameter
-import io.github.pshevche.spockk.compilation.ir.irImplicitNotNull
 import io.github.pshevche.spockk.compilation.ir.irTry
-import io.github.pshevche.spockk.compilation.ir.irType
-import io.github.pshevche.spockk.compilation.ir.irVal
 import io.github.pshevche.spockk.compilation.ir.isAssertCall
 import io.github.pshevche.spockk.compilation.ir.requiredThisParameter
 import io.github.pshevche.spockk.compilation.ir.unwrapImplicitCoercionToUnit
 import io.github.pshevche.spockk.compilation.transformer.InternalIdentifiers.CONDITION_THROWABLE_VAR
-import io.github.pshevche.spockk.compilation.transformer.InternalIdentifiers.ERROR_COLLECTOR_VAR
-import io.github.pshevche.spockk.compilation.transformer.InternalIdentifiers.VALUE_RECORDER_VAR
 import io.github.pshevche.spockk.compilation.transformer.SpockkIrRewriter
 import io.github.pshevche.spockk.compilation.transformer.ir.IrValueRecorder
 import io.github.pshevche.spockk.compilation.transformer.ir.SpockkIrRewriterContext
 import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
 import org.jetbrains.kotlin.backend.common.lower.irCatch
+import org.jetbrains.kotlin.ir.IrBuiltIns
 import org.jetbrains.kotlin.ir.IrStatement
 import org.jetbrains.kotlin.ir.builders.irBlock
-import org.jetbrains.kotlin.ir.builders.irCallConstructor
-import org.jetbrains.kotlin.ir.builders.irGetField
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrVariable
 import org.jetbrains.kotlin.ir.expressions.IrCall
 import org.jetbrains.kotlin.ir.expressions.IrCatch
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.symbols.UnsafeDuringIrConstructionAPI
-import org.jetbrains.kotlin.ir.util.constructors
 import org.jetbrains.kotlin.ir.util.file
 
 @OptIn(UnsafeDuringIrConstructionAPI::class)
@@ -51,40 +41,28 @@ internal class ConditionRewriter(
   override val rewriterContext: SpockkIrRewriterContext,
   private val builder: DeclarationIrBuilder,
   private val feature: IrFunction,
-  private val blockOrdinal: Int
+  private val blockOrdinal: Int,
+  private val valueRecorderVar: IrVariable?,
+  private val errorCollectorVar: IrVariable?
 ) : SpockkIrRewriter {
 
   fun rewrite(statements: List<IrStatement>): List<IrStatement> = buildList {
-    val hasConditionStatements = statements.any { isConditionStatement(it) }
-
-    if (hasConditionStatements) {
-      val valueRecorderVar = initializeValueRecorderStatement().also { add(it) }
-      val errorCollectorVar = initializeErrorCollectorStatement().also { add(it) }
-      add(
-        rewriterContext.spockRuntime.irCallBlockEntered(
-          builder,
-          feature.requiredThisParameter(),
-          blockOrdinal
-        )
+    add(
+      rewriterContext.spockRuntime.irCallBlockEntered(
+        builder,
+        feature.requiredThisParameter(),
+        blockOrdinal
       )
-      statements.forEach {
-        if (isConditionStatement(it)) {
-          add(rewriteConditionStatement(it as IrExpression, valueRecorderVar, errorCollectorVar))
-        } else {
-          add(it)
-        }
+    )
+    statements.forEach { statement ->
+      if (statement.isConditionStatement(irBuiltIns)) {
+        // A condition statement implies the feature has conditions, so the shared value recorder
+        // and error collector were created once at feature scope (see FeatureRewriter).
+        add(rewriteConditionStatement(statement as IrExpression, valueRecorderVar!!, errorCollectorVar!!))
+      } else {
+        add(statement)
       }
-    } else {
-      add(
-        rewriterContext.spockRuntime.irCallBlockEntered(
-          builder,
-          feature.requiredThisParameter(),
-          blockOrdinal
-        )
-      )
-      addAll(statements)
     }
-
     add(
       rewriterContext.spockRuntime.irCallBlockExited(
         builder,
@@ -92,39 +70,6 @@ internal class ConditionRewriter(
         blockOrdinal
       )
     )
-  }
-
-  private fun isConditionStatement(statement: IrStatement): Boolean {
-    val expr = (statement as? IrExpression)?.unwrapImplicitCoercionToUnit() ?: return false
-    return expr.isAssertCall() || hasBooleanReturnType(expr)
-  }
-
-  private fun hasBooleanReturnType(statement: IrExpression): Boolean = statement.type == irBuiltIns.booleanType
-
-  private fun initializeErrorCollectorStatement(): IrVariable {
-    val errorCollectorType = builder.irType(IrIdentifiers.Spock.ERROR_COLLECTOR_FQN)
-    val errorRethrowerType = builder.irType(IrIdentifiers.Spock.ERROR_RETHROWER_FQN)
-    val errorRethrowerClass = rewriterContext.findRequiredClassSymbol(IrIdentifiers.Spock.ERROR_RETHROWER_FQN)
-    val instanceField = errorRethrowerClass.findFieldByName("INSTANCE")
-
-    val instanceFieldAccess = builder.irGetField(null, instanceField).apply {
-      superQualifierSymbol = errorRethrowerClass
-    }
-    return irVal(ERROR_COLLECTOR_VAR, errorCollectorType).apply {
-      parent = feature
-      initializer = builder.irImplicitNotNull(instanceFieldAccess, errorRethrowerType)
-    }
-  }
-
-  private fun initializeValueRecorderStatement(): IrVariable {
-    val valueRecorderClass = rewriterContext.findRequiredClassSymbol(IrIdentifiers.Spock.VALUE_RECORDER_FQN)
-    val valueRecorderConstructor = valueRecorderClass.constructors.first()
-    val newValueRecorderCall = builder.irCallConstructor(valueRecorderConstructor, listOf())
-
-    return irVal(VALUE_RECORDER_VAR, builder.irType(IrIdentifiers.Spock.VALUE_RECORDER_FQN)).apply {
-      parent = feature
-      initializer = newValueRecorderCall
-    }
   }
 
   private fun rewriteConditionStatement(
@@ -183,4 +128,14 @@ internal class ConditionRewriter(
       irCatch(catchVar, irBlock { +catchResult })
     }
   }
+}
+
+/**
+ * A statement is a condition when it is an explicit `assert(...)` call or an implicit boolean
+ * expression statement. Shared by [ConditionRewriter] and the feature rewriter, which decides up
+ * front whether a feature needs a value recorder.
+ */
+internal fun IrStatement.isConditionStatement(irBuiltIns: IrBuiltIns): Boolean {
+  val expr = (this as? IrExpression)?.unwrapImplicitCoercionToUnit() ?: return false
+  return expr.isAssertCall() || expr.type == irBuiltIns.booleanType
 }
